bot1/Chatbot/requirements.txt
langchain_openai # For using OpenAI API
langchain_core # For embedding chatprompt template
python_dotenv
streamlit # For UI
langchain_community # For calling open source LLMs
langserve # For deployments, creating API routes
langchain_groq